introduce a "scheduler" trait that can be set with `in_scheduler`
The scheduler gets threaded down through the calls to `drive()` and
`with_producer()`. It replaces the existing "bridge" functions. This
version includes two schedulers -- the adaptive scheduler and the
fixed-length scheduler.

(Note that the existing `min_len()` and `max_len()` combinators are
ignored by this commit; they will be replaced by uses of
`in_scheduler()` in the next commit.)